fix(autoware_tensorrt_e2e): compose the SE(2) warp transform in double, once
The warp kernel lifted every output cell into the world frame in float32,
with poses already stored as float. Map coordinates are ~1e5 m on T4-style
maps: the float pose quantizes slow-speed inter-frame displacements to
centimetres, and the per-cell world-frame round trip cancels two ~1e5 m
numbers to recover a metre-scale offset 32,400 times per frame. The upstream
reference (deployment/temporal.py) composes T_source^-1 . T_current once in
float64 for exactly this reason.

Do the same: poses stay double end to end (odometry -> cache slot), the host
composes the relative rotation and metre-scale shift in double, and the
kernel only ever sees metre-scale float32 values.

New golden test replays cases generated by the reference implementation
itself (OnePlanner projects/resworld/scripts/gen_bev_warp_goldens.py) through
the CUDA kernel. Before this change the UTM-scale cases failed at max|err|
2.2e-3 / 3.4e-3 against unit-scale features; after it the worst case is 9e-6
(gate 1e-4). 22/22 tests pass; synthetic 0.2 s-contract pipeline re-verified
at 10.00 Hz.

+++ e2e/autoware_tensorrt_e2e/include/autoware/tensorrt_e2e/bev_feature/temporal_bev_cache.hpp
@@ -75,13 +75,16 @@ public:
   /**
    * @brief Insert the newest feature map (device `[C, H, W]`) with its source ego pose.
    *
-   * The pose is `[x, y, cos(yaw), sin(yaw)]` in the map frame. A stamp at or before the newest
-   * cached stamp resets the cache first and reports `kGapReset`. Maps that fall out of the
-   * history window are evicted (their device buffers are recycled). Synchronizes `stream`
-   * before returning (the source buffer may be reused by the caller afterwards).
+   * The pose is `[x, y, cos(yaw), sin(yaw)]` in the map frame, kept in double: map coordinates
+   * are ~1e5 m on T4-style maps, and the SE(2) warp composes pose differences — float storage
+   * would quantize a slow-speed inter-frame displacement to centimetres before the warp ever
+   * sees it. A stamp at or before the newest cached stamp resets the cache first and reports
+   * `kGapReset`. Maps that fall out of the history window are evicted (their device buffers
+   * are recycled). Synchronizes `stream` before returning (the source buffer may be reused by
+   * the caller afterwards).
    */
   InsertResult insert(
-    const float * d_feature, const std::array<float, 4> & pose, const rclcpp::Time & stamp,
+    const float * d_feature, const std::array<double, 4> & pose, const rclcpp::Time & stamp,
     cudaStream_t stream);
 
   bool ready() const;
@@ -113,7 +116,7 @@ private:
   struct Slot
   {
     autoware::cuda_utils::CudaUniquePtr<float[]> feature;
-    std::array<float, 4> pose{};
+    std::array<double, 4> pose{};
     rclcpp::Time stamp;
   };
